requirements.txt
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.37.0
Pillow>=10.0.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.8.0
discord.py>=2.3.2
python-dotenv>=1.0.0
flask==3.0.0
plotly==5.18.0
datasets>=2.16.0
nltk>=3.8.1
spacy>=3.7.2
gensim>=4.3.2
pytest>=7.4.0
black>=23.12.0
mypy>=1.8.0
pylint>=3.0.0
pre-commit>=3.6.0
mlflow>=2.9.0
optuna>=3.5.0
shap>=0.44.0
tensorboard>=2.15.0
lime>=0.2.0.1
requests>=2.31.0
gdown>=4.7.0
twitchio>=2.8.0 